etl-worldbank-dashboard-base-de-datos/requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.8.0
narwhals==2.17.0
numpy==1.26.4
openpyxl==3.1.2
packaging==26.0
pandas==2.2.2
pillow==12.1.1
plotly==6.5.2
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
schedule==1.2.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.48
streamlit==1.54.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
